Adds non-ASCII characters to test suite
There's a bug where non-ASCII filenames and directory names cause the
resulting `mp3` file to be transcoded incorrectly (invalid). This test
verifies that the issue is not due to filename comparisons.

It may have to do with the FAT32 filesystem that is used as a
destination. More testing is needed.